requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.9
supabase>=2.3.4
pandas>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
passlib>=1.7.4
openpyxl>=3.1.2
